backend/requirements.txt
aiofiles==23.1.0
amqp==5.3.1
ansible==13.0.0
ansible-core==2.20.0
anyio==4.11.0
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
cbor2==5.8.0
celery==5.6.2
certifi==2025.4.26
cffi==2.0.0
chai==1.1.2
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.2
cryptography==46.0.5
daphne==4.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
Django==4.2.16
django-allauth==65.9.0
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-extensions==4.1
django-filter==23.5
django-storages==1.14.6
django-taggit==6.1.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djongo==1.3.6
djoser==2.3.1
dnspython==2.8.0
drf-nested-routers==0.95.0
drf-yasg==1.21.10
fastapi==0.95.2
filelock==3.19.1
flake8==6.1.0
Flask==2.3.2
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.6
jiter==0.13.0
kombu==5.6.2
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.2
numpy==2.3.2
oauthlib==3.2.2
openai==2.21.0
packaging==25.0
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==1.10.11
pyflakes==3.1.0
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.15.3
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==7.3.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==7.2.0
requests==2.31.0
requests-oauthlib==2.0.0
resolvelib==1.2.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.8.3
sqlparse==0.5.1
starlette==0.27.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.22.0
vine==5.1.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
zope.interface==8.2
